requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pillow>=10.0.0
supabase>=2.0.0
fpdf2>=2.7.0
voyageai>=0.3.0
httpx>=0.27.0
